feat: move WorksList to separate component and remove maps.tsx page
